isisd: move old BFD command to fabricd only
The `isisd` will receive a northbound version of the BFD command, so
this is the first step to implement it.
